refactored sdk package
