+++ android/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.customerioreactnative">
+          package="io.customer.reactnative.sdk">
 
 </manifest>
