+++ android/src/main/java/io/customer/reactnative/sdk/CustomerioReactnativePackage.kt
@@ -1,4 +1,4 @@
-package com.customerioreactnative
+package io.customer.reactnative.sdk
 
 import com.facebook.react.ReactPackage
 import com.facebook.react.bridge.NativeModule
